backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Async & WebSocket Support
websockets==12.0
python-socketio==5.11.0
aiohttp==3.9.1

# LangChain & LangGraph - Updated for Python 3.13 compatibility
langchain>=0.1.0
langchain-openai>=0.0.8
langgraph>=0.0.20
langchain-community>=0.0.20

# Vector Database & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1
tiktoken>=0.5.2  # Compatible with langchain-openai

# GitHub Integration
PyGithub==2.1.1
gitpython==3.1.41

# Environment & Configuration
python-dotenv==1.0.0
pydantic>=2.7.0  # Python 3.13 compatible
pydantic-settings==2.1.0

# Security
cryptography==41.0.7
pyjwt==2.8.0
python-jose[cryptography]==3.3.0

# HTTP & API
httpx>=0.26.0
requests==2.31.0

# Data Processing
pandas>=2.2.0  # Python 3.13 compatible
numpy>=1.26.0,<2.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Caching & Queue
redis==5.0.1
celery==5.3.6

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-httpx==0.30.0  # Updated package name

# Code Quality
black==23.12.1
ruff==0.1.11
mypy==1.8.0

# Utilities
tenacity==8.2.3
python-dateutil==2.8.2
pytz==2023.3.post1
